Clean code

+++ simulation/camera/lane.py
@@ -22,6 +22,7 @@ class LaneDetection(threading.Thread):
         self.rawCapture = None
         self.current_center_list = []
         self.measurements = measurements
+        self.perform_ld = True
 
     def init_camera(self):
         self.camera = picamera.PiCamera()
@@ -187,8 +188,8 @@ class LaneDetection(threading.Thread):
         """
         if len(lines) > 0:
             left_lane, right_lane = self.average_slope(lines)
-	    logging.info("LEFT LANE: " + str(left_lane)
-	    #logging.info("RIGHT LANE: " + str(right_lane)           y1 = image.shape[0]
+            logging.info("LEFT LANE: " + str(left_lane)
+            #logging.info("RIGHT LANE: " + str(right_lane)           y1 = image.shape[0]
             y2 = y1 * 0.6
             left_line = self.pixel_points(y1, y2, left_lane)
             right_line = self.pixel_points(y1, y2, right_lane)
@@ -230,46 +231,48 @@ class LaneDetection(threading.Thread):
 
         while not self.exitFlag:
             for frame in self.camera.capture_continuous(self.rawCapture, format="bgr", use_video_port=True):
-                with open("calibration_matrix.txt", 'r') as f:
-                    calibration_matrix = pickle.load(f)
+                if self.perform_ld:
+                    with open("calibration_matrix.txt", 'r') as f:
+                        calibration_matrix = pickle.load(f)
 
-                img = frame.array
+                    img = frame.array
+                    undistorted = cv2.undistort(img, calibration_matrix['mtx'], calibration_matrix['dist'], None,
+                                                calibration_matrix['mtx'])
+                    gray = self.gray_scale(img)
+                    blurred = self.gaussian_smoothing(gray, 3)
+                    edged = self.canny_detector(blurred, 50, 150)
+                    trimmed = self.region_selection(edged)
+                    houghlines = self.hough_transform(trimmed)
 
-                undistorted = cv2.undistort(img, calibration_matrix['mtx'], calibration_matrix['dist'], None,
-                                            calibration_matrix['mtx'])
+                    if self.debug:
+                        final = self.draw_lane_lines(img, self.lane_lines(img, houghlines))
+                        cv2.line(final, (0, 450), (230, 300), [0, 255, 0], 2)
+                        cv2.line(final, (720, 480), (540, 310), [0, 255, 0], 2)
+                        cv2.line(final, (230, 300), (540, 310), [0, 255, 0], 2)
+                        cv2.imshow("Lane Detection", final)
 
-                gray = self.gray_scale(img)
-                blurred = self.gaussian_smoothing(gray, 3)
-                edged = self.canny_detector(blurred, 50, 150)
-                trimmed = self.region_selection(edged)
-                houghlines = self.hough_transform(trimmed)
+                        key = cv2.waitKey(1) & 0xFF
+                        self.rawCapture.truncate()
+                        self.rawCapture.seek(0)
+                        if key == ord("q"):
+                            break
+                    else:
+                        current_center = self.draw_lane_lines(undistorted, self.lane_lines(img, houghlines))
+                        offset = settings.ACTUAL_CENTER - current_center
+                        self.rawCapture.truncate()
+                        self.rawCapture.seek(0)
 
-                if self.debug:
-                    final = self.draw_lane_lines(img, self.lane_lines(img, houghlines))
-                    cv2.line(final, (0, 450), (230, 300), [0, 255, 0], 2)
-                    cv2.line(final, (720, 480), (540, 310), [0, 255, 0], 2)
-                    cv2.line(final, (230, 300), (540, 310), [0, 255, 0], 2)
-                    cv2.imshow("Lane Detection", final)
+                        logging.info("Adding new measurement to list - " + str(current_center))
+                        self.current_center_list.append(current_center)
 
-                    key = cv2.waitKey(1) & 0xFF
-                    self.rawCapture.truncate()
-                    self.rawCapture.seek(0)
-                    if key == ord("q"):
-                        break
-                else:
-                    current_center = self.draw_lane_lines(undistorted, self.lane_lines(img, houghlines))
-                    offset = settings.ACTUAL_CENTER - current_center
-                    self.rawCapture.truncate()
-                    self.rawCapture.seek(0)
-
-                    logging.info("Adding new measurement to list - " + str(current_center))
-                    self.current_center_list.append(current_center)
-
-                    if len(self.current_center_list) % 3 == 0:
-                        logging.info("Adding new measurements to queue - " + str(self.current_center_list[-5:]))
-                        self.measurements.put(self.current_center_list[-5:])
-                    # return current_center, offset
+                        if len(self.current_center_list) % 3 == 0:
+                            logging.info("Adding new measurements to queue - " + str(self.current_center_list[-5:]))
+                            self.measurements.put(self.current_center_list[-5:])
+                        # return current_center, offset
 
     def stop_thread(self):
         self.exitFlag = True
         self.camera.close()
+
+    def stop_ld(self, value=True):
+        self.perform_ld = value
